refactor(agent-gui): separate final text projection
