+++ packages/agent/gui/shared/agentConversation/projection/agentConversationProjection.ts
@@ -11,6 +11,7 @@ import type {
 } from "../contracts/agentMessageRowVM";
 import type { AgentToolCallVM } from "../contracts/agentToolCallVM";
 import type { AgentTranscriptRowVM } from "../contracts/agentTranscriptRowVM";
+import { projectAgentMessageFinalText } from "./agentMessageFinalTextProjection";
 import { computeAgentToolGroups } from "./agentToolGroupingProjection";
 import { buildAgentTurnSequenceItems } from "./agentTurnSequenceProjection";
 import { projectTurnRows } from "./agentTurnRowProjection";
@@ -58,7 +59,7 @@ export function projectAgentConversationVM(
     rows.push(processing);
   }
 
-  const normalizedRows = projectMessageCopyText(
+  const normalizedRows = projectAgentMessageFinalText(
     dropCodexRuntimeDiagnosticNotices(
       dropRedundantErrorWarningNotices(
         mergeAdjacentAssistantMessageRows(
@@ -68,9 +69,7 @@ export function projectAgentConversationVM(
         )
       )
     ),
-    {
-      assistantCopyEligibleTurnIds: buildAssistantCopyEligibleTurnIds(detail)
-    }
+    detail
   );
 
   return {
@@ -415,164 +414,6 @@ function mergeSourceTimelineItems(
     return previous;
   }
   return [...previous, ...next];
-}
-
-function projectMessageCopyText(
-  rows: readonly AgentTranscriptRowVM[],
-  options: { assistantCopyEligibleTurnIds: ReadonlySet<string> }
-): AgentTranscriptRowVM[] {
-  const assistantCopyTargetKeys = findLatestAssistantCopyTargetKeys(
-    rows,
-    options.assistantCopyEligibleTurnIds
-  );
-  return rows.map((row) => {
-    if (row.kind !== "message") {
-      return row;
-    }
-
-    let changed = false;
-    const messages = row.messages.map((message) => {
-      const isTurnFinalText =
-        row.speaker === "assistant" &&
-        assistantCopyTargetKeys.has(messageCopyTargetKey(row, message));
-      const copyText =
-        row.speaker === "user"
-          ? copyTextForUserMessage(message)
-          : isTurnFinalText
-            ? message.body
-            : null;
-      if (
-        (message.copyText ?? null) === copyText &&
-        message.isTurnFinalText === (isTurnFinalText ? true : undefined)
-      ) {
-        return message;
-      }
-      changed = true;
-      const nextMessage = isTurnFinalText
-        ? { ...message, isTurnFinalText: true as const }
-        : omitTurnFinalText(message);
-      if (copyText) {
-        return { ...nextMessage, copyText };
-      }
-      const { copyText: _copyText, ...withoutCopyText } = nextMessage;
-      return withoutCopyText;
-    });
-
-    return changed ? { ...row, messages } : row;
-  });
-}
-
-function omitTurnFinalText(
-  message: AgentMessageContentVM
-): AgentMessageContentVM {
-  const { isTurnFinalText: _isTurnFinalText, ...withoutTurnFinalText } =
-    message;
-  return withoutTurnFinalText;
-}
-
-function buildAssistantCopyEligibleTurnIds(
-  detail: WorkspaceAgentSessionDetailViewModel
-): ReadonlySet<string> {
-  const ids = new Set<string>();
-  detail.turns.forEach((turn, index) => {
-    if (
-      index < detail.turns.length - 1 ||
-      isLatestTranscriptTurnSettled(detail)
-    ) {
-      ids.add(turn.id);
-    }
-  });
-  return ids;
-}
-
-function isLatestTranscriptTurnSettled(
-  detail: WorkspaceAgentSessionDetailViewModel
-): boolean {
-  const latestTranscriptTurnId = detail.turns.at(-1)?.id;
-  const canonicalTurn = detail.sessionTurns?.find(
-    (turn) => turn.turnId === latestTranscriptTurnId
-  );
-  const activeTurn = detail.session.activeTurn;
-  if (
-    activeTurn &&
-    activeTurn.turnId === latestTranscriptTurnId &&
-    activeTurn.phase !== "settled"
-  ) {
-    return false;
-  }
-  if (canonicalTurn) {
-    return (
-      canonicalTurn.phase === "settled" &&
-      detail.showProcessingIndicator !== true
-    );
-  }
-  const activePhase = activeTurn?.phase ?? "";
-  return (
-    detail.showProcessingIndicator !== true &&
-    !["submitted", "running", "waiting", "settling"].includes(activePhase)
-  );
-}
-
-function findLatestAssistantCopyTargetKeys(
-  rows: readonly AgentTranscriptRowVM[],
-  eligibleTurnIds: ReadonlySet<string>
-): ReadonlySet<string> {
-  const targetKeys = new Set<string>();
-  const coveredTurnIds = new Set<string>();
-  for (let rowIndex = rows.length - 1; rowIndex >= 0; rowIndex -= 1) {
-    const row = rows[rowIndex];
-    if (
-      row?.kind !== "message" ||
-      row.speaker !== "assistant" ||
-      coveredTurnIds.has(row.turnId) ||
-      !eligibleTurnIds.has(row.turnId)
-    ) {
-      continue;
-    }
-    for (
-      let messageIndex = row.messages.length - 1;
-      messageIndex >= 0;
-      messageIndex -= 1
-    ) {
-      const message = row.messages[messageIndex];
-      if (message && isSettledTextMessageCopyCandidate(message)) {
-        targetKeys.add(messageCopyTargetKey(row, message));
-        coveredTurnIds.add(row.turnId);
-        break;
-      }
-    }
-  }
-  return targetKeys;
-}
-
-function messageCopyTargetKey(
-  row: AgentMessageRowVM,
-  message: AgentMessageContentVM
-): string {
-  return `${row.id}\u0000${message.id}`;
-}
-
-function copyTextForUserMessage(message: AgentMessageContentVM): string | null {
-  return isTextMessageCopyCandidate(message) ? message.body : null;
-}
-
-function isSettledTextMessageCopyCandidate(
-  message: AgentMessageContentVM
-): boolean {
-  return (
-    isTextMessageCopyCandidate(message) &&
-    message.statusKind !== "working" &&
-    message.statusKind !== "waiting"
-  );
-}
-
-function isTextMessageCopyCandidate(message: AgentMessageContentVM): boolean {
-  return (
-    message.body.trim() !== "" &&
-    message.contentKind !== "image-grid" &&
-    !message.visibleError &&
-    !message.systemNotice
-  );
 }
 
 function isMergeableAssistantMessageRow(
